altered price filed to floatfield

+++ client/models.py
@@ -24,7 +24,7 @@ class Task(models.Model):
     expiry_date = models.DateField(auto_now = False)
     documet_format = models.CharField(max_length=100,choices=FORMAT_AS_CHOICES, default='None')
     task_file = models.FileField(upload_to='Tasks/', max_length=150, null=True, blank=True)
-    price = models.DecimalField(max_digits=7, decimal_places=2)
+    price = models.FloatField()
     is_taken = models.BooleanField(default=False)
     show = models.BooleanField(default=True)
     paid = models.BooleanField(default=False)
